Removed ssm subdir from templates, moved fileIconPath functions to vsCodeUtils
