--- src/awsexplorer/utils.ts
@@ -1,33 +1,0 @@
-/*!
- * Copyright 2018 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- * SPDX-License-Identifier: Apache-2.0
- */
-import * as vscode from 'vscode'
-import { isFileIconThemeSeti } from '../shared/utilities/vsCodeUtils'
-import { ext } from '../shared/extensionGlobals'
-
-export function fileIconPath(): vscode.ThemeIcon | { light: vscode.Uri; dark: vscode.Uri } {
-    // Workaround for https://github.com/microsoft/vscode/issues/85654
-    // Once this is resolved, ThemeIcons can be used for seti as well
-    if (isFileIconThemeSeti()) {
-        return {
-            dark: vscode.Uri.file(ext.iconPaths.dark.file),
-            light: vscode.Uri.file(ext.iconPaths.light.file),
-        }
-    } else {
-        return vscode.ThemeIcon.File
-    }
-}
-
-export function folderIconPath(): vscode.ThemeIcon | { light: vscode.Uri; dark: vscode.Uri } {
-    // Workaround for https://github.com/microsoft/vscode/issues/85654
-    // Once this is resolved, ThemeIcons can be used for seti as well
-    if (isFileIconThemeSeti()) {
-        return {
-            dark: vscode.Uri.file(ext.iconPaths.dark.folder),
-            light: vscode.Uri.file(ext.iconPaths.light.folder),
-        }
-    } else {
-        return vscode.ThemeIcon.Folder
-    }
-}
