+++ src/s3/explorer/s3FolderNode.ts
@@ -21,7 +21,7 @@ import { makeChildrenNodes } from '../../shared/treeview/treeNodeUtilities'
 import { localize } from '../../shared/utilities/vsCodeUtils'
 import { ChildNodeLoader } from '../../awsexplorer/childNodeLoader'
 import { ChildNodePage } from '../../awsexplorer/childNodeLoader'
-import { folderIconPath } from '../../awsexplorer/utils'
+import { folderIconPath } from '../../shared/utilities/vsCodeUtils'
 import { S3FileNode } from './s3FileNode'
 import { inspect } from 'util'
 import { Workspace } from '../../shared/vscode/workspace'
