+++ src/shared/utilities/vsCodeUtils.ts
@@ -7,6 +7,7 @@ import * as vscode from 'vscode'
 import * as nls from 'vscode-nls'
 import { getLogger, LogLevel } from '../logger'
 import { Loggable } from '../logger/loggableType'
+import { ext } from '../../shared/extensionGlobals'
 
 // TODO: Consider NLS initialization/configuration here & have packages to import localize from here
 export const localize = nls.loadMessageBundle()
@@ -120,4 +121,30 @@ export function getChannelLogger(channel: vscode.OutputChannel): ChannelLogger {
 export function isFileIconThemeSeti(): boolean {
     const iconTheme = vscode.workspace.getConfiguration('workbench').get('iconTheme')
     return !iconTheme || iconTheme === 'vs-seti'
+}
+
+export function fileIconPath(): vscode.ThemeIcon | { light: vscode.Uri; dark: vscode.Uri } {
+    // Workaround for https://github.com/microsoft/vscode/issues/85654
+    // Once this is resolved, ThemeIcons can be used for seti as well
+    if (isFileIconThemeSeti()) {
+        return {
+            dark: vscode.Uri.file(ext.iconPaths.dark.file),
+            light: vscode.Uri.file(ext.iconPaths.light.file),
+        }
+    } else {
+        return vscode.ThemeIcon.File
+    }
+}
+
+export function folderIconPath(): vscode.ThemeIcon | { light: vscode.Uri; dark: vscode.Uri } {
+    // Workaround for https://github.com/microsoft/vscode/issues/85654
+    // Once this is resolved, ThemeIcons can be used for seti as well
+    if (isFileIconThemeSeti()) {
+        return {
+            dark: vscode.Uri.file(ext.iconPaths.dark.folder),
+            light: vscode.Uri.file(ext.iconPaths.light.folder),
+        }
+    } else {
+        return vscode.ThemeIcon.Folder
+    }
 }
