+++ templates/HelloWorldPython.ssm.yaml
